gym_power/requirements.txt
django
psycopg2-binary
psycopg2
djangorestframework
django-filter
reportlab
pandas
openpyxl
requests
numpy